+++ conan/tools/apple/apple.py
@@ -54,78 +54,3 @@ def apple_min_version_flag(os_version, os_sdk, subsystem):
     if not flag:
         return ''
     return "%s=%s" % (flag, os_version)
-
-
-class XCRun(object):
-
-    def __init__(self, settings, sdk=None):
-        """sdk=False will skip the flag
-           sdk=None will try to adjust it automatically"""
-        if sdk is None and settings:
-            sdk = settings.get_safe('os.sdk')
-        self.sdk = sdk
-
-    def _invoke(self, args):
-        def cmd_output(cmd):
-            from conans.util.runners import check_output_runner
-            return check_output_runner(cmd).strip()
-
-        command = ['xcrun']
-        if self.sdk:
-            command.extend(['-sdk', self.sdk])
-        command.extend(args)
-        return cmd_output(command)
-
-    def find(self, tool):
-        """find SDK tools (e.g. clang, ar, ranlib, lipo, codesign, etc.)"""
-        return self._invoke(['--find', tool])
-
-    @property
-    def sdk_path(self):
-        """obtain sdk path (aka apple sysroot or -isysroot"""
-        return self._invoke(['--show-sdk-path'])
-
-    @property
-    def sdk_version(self):
-        """obtain sdk version"""
-        return self._invoke(['--show-sdk-version'])
-
-    @property
-    def sdk_platform_path(self):
-        """obtain sdk platform path"""
-        return self._invoke(['--show-sdk-platform-path'])
-
-    @property
-    def sdk_platform_version(self):
-        """obtain sdk platform version"""
-        return self._invoke(['--show-sdk-platform-version'])
-
-    @property
-    def cc(self):
-        """path to C compiler (CC)"""
-        return self.find('clang')
-
-    @property
-    def cxx(self):
-        """path to C++ compiler (CXX)"""
-        return self.find('clang++')
-
-    @property
-    def ar(self):
-        """path to archiver (AR)"""
-        return self.find('ar')
-
-    @property
-    def ranlib(self):
-        """path to archive indexer (RANLIB)"""
-        return self.find('ranlib')
-
-    @property
-    def strip(self):
-        """path to symbol removal utility (STRIP)"""
-        return self.find('strip')
-
-    @property
-    def libtool(self):
-        """path to libtool"""
-        return self.find('libtool')
